Updating reference images... I have thought about this... but wanted to commit what changed before I lose this in my container. Updating the test_mrfgen.py script to hvae a flag for saving the output.  Also changed one of the JPEG tests. One test is still failing.  Need to investigate. At first glance, it was also failing on-prem in the 1.3.7 baseline... so ... TBD

+++ src/test/test_mrfgen.py
@@ -275,7 +275,10 @@ class TestMRFGeneration_nonpaletted(unittest.TestCase):
         mrf = None
 
     def tearDown(self):
-        shutil.rmtree(self.staging_area)
+        if not SAVE_RESULTS:
+            shutil.rmtree(self.staging_area)
+        else:
+            print "Leaving test results in : " + self.staging_area
         
 
 class TestMRFGeneration_polar(unittest.TestCase):
@@ -364,11 +367,15 @@ class TestMRFGeneration_polar(unittest.TestCase):
             print "Comparing: " + self.output_img + " to " + self.compare_img
         self.assertTrue(filecmp.cmp(self.output_img, self.compare_img), "Output image does not match")
 
+
         img = None
         mrf = None
         
     def tearDown(self):
-        shutil.rmtree(self.staging_area)
+        if not SAVE_RESULTS:
+            shutil.rmtree(self.staging_area)
+        else:
+            print "Leaving test results in : " + self.staging_area
         
 
 class TestMRFGeneration_mercator(unittest.TestCase):
@@ -463,7 +470,10 @@ class TestMRFGeneration_mercator(unittest.TestCase):
         mrf = None
 
     def tearDown(self):
-        shutil.rmtree(self.staging_area)
+        if not SAVE_RESULTS:
+            shutil.rmtree(self.staging_area)
+        else:
+            print "Leaving test results in : " + self.staging_area
         
 class TestMRFGeneration_OBPG(unittest.TestCase):
     
@@ -701,8 +711,11 @@ class TestMRFGeneration_OBPG(unittest.TestCase):
             con.close()
 
     def tearDown(self):
-        [os.remove(os.path.join(self.input_dir, file)) for file in os.listdir(self.input_dir) if not file.endswith('.tiff')]
-        shutil.rmtree(self.staging_area)
+        if not SAVE_RESULTS:
+            [os.remove(os.path.join(self.input_dir, file)) for file in os.listdir(self.input_dir) if not file.endswith('.tiff')]
+            shutil.rmtree(self.staging_area)
+        else:
+            print "Leaving test results in : " + self.staging_area
         
 class TestMRFGeneration_OBPG_webmerc(unittest.TestCase):
     
@@ -821,8 +834,11 @@ class TestMRFGeneration_OBPG_webmerc(unittest.TestCase):
             con.close()
 
     def tearDown(self):
-        [os.remove(os.path.join(self.input_dir, file)) for file in os.listdir(self.input_dir) if not file.endswith('.tiff')]
-        shutil.rmtree(self.staging_area)
+        if not SAVE_RESULTS:
+            [os.remove(os.path.join(self.input_dir, file)) for file in os.listdir(self.input_dir) if not file.endswith('.tiff')]
+            shutil.rmtree(self.staging_area)
+        else:
+            print "Leaving test results in : " + self.staging_area
 
 class TestMRFGeneration_tiled_z(unittest.TestCase):
     '''
@@ -936,7 +952,10 @@ class TestMRFGeneration_tiled_z(unittest.TestCase):
             con.close()
 
     def tearDown(self):
-        shutil.rmtree(self.staging_area)
+        if not SAVE_RESULTS:
+            shutil.rmtree(self.staging_area)
+        else:
+            print "Leaving test results in : " + self.staging_area
 
 
 class TestMRFGeneration_nonpaletted_colormap(unittest.TestCase):
@@ -1021,7 +1040,10 @@ class TestMRFGeneration_nonpaletted_colormap(unittest.TestCase):
         mrf = None
 
     def tearDown(self):
-        shutil.rmtree(self.staging_area)
+        if not SAVE_RESULTS:
+            shutil.rmtree(self.staging_area)
+        else:
+            print "Leaving test results in : " + self.staging_area
 
 class TestMRFGeneration_email_notification(unittest.TestCase):
 
@@ -1097,6 +1119,7 @@ class TestMRFGeneration_email_notification(unittest.TestCase):
 
     def tearDown(self):
         shutil.rmtree(self.staging_area)
+
 
 class TestMRFGeneration_mixed_projections(unittest.TestCase):
     
@@ -1189,7 +1212,10 @@ class TestMRFGeneration_mixed_projections(unittest.TestCase):
         mrf = None
         
     def tearDown(self):
-        shutil.rmtree(self.staging_area)
+        if not SAVE_RESULTS:
+            shutil.rmtree(self.staging_area)
+        else:
+            print "Leaving test results in : " + self.staging_area
 
 class TestMRFGeneration_antimeridian_crossing(unittest.TestCase):
     
@@ -1282,7 +1308,10 @@ class TestMRFGeneration_antimeridian_crossing(unittest.TestCase):
         mrf = None
         
     def tearDown(self):
-        shutil.rmtree(self.staging_area)        
+        if not SAVE_RESULTS:
+            shutil.rmtree(self.staging_area)
+        else:
+            print "Leaving test results in : " + self.staging_area
 
 if __name__ == '__main__':
     # Parse options before running tests
